requirements.txt
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.3.0
azure-identity==1.4.0
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coverage==5.0
docutils==0.16
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.0.2
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pluggy==0.13.1
py==1.8.0
Pygments==2.5.2
pyparsing==2.4.5
pytest==5.2.3
pytest-cov==2.8.1
pytest-mock==3.5.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-mock==1.7.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0